Restructure the datastore package.
