+++ pkg/datastore/inmemory/datastore_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package datastore
+package inmemory
 
 import (
 	"fmt"
@@ -30,7 +30,7 @@ func (t testValue) Clone() datalayer.Cloneable { return testValue{Value: t.Value
 
 // TestGetOrCreateModel tests creation, same-instance return, and attribute persistence.
 func TestGetOrCreateModel(t *testing.T) {
-	s := NewInMemoryStore()
+	s := NewDatastore()
 
 	m := s.GetOrCreateModel("llama-3")
 	if m == nil || m.GetName() != "llama-3" || m.GetAttributes() == nil {
@@ -49,7 +49,7 @@ func TestGetOrCreateModel(t *testing.T) {
 
 // TestDeleteModel tests that delete+recreate yields a fresh model and that deleting a missing key is a no-op.
 func TestDeleteModel(t *testing.T) {
-	s := NewInMemoryStore()
+	s := NewDatastore()
 
 	s.GetOrCreateModel("llama-3").GetAttributes().Put("key", testValue{Value: 1})
 	s.DeleteModel("llama-3")
@@ -62,7 +62,7 @@ func TestDeleteModel(t *testing.T) {
 
 // TestModelsIsolated tests that different models have independent Attributes.
 func TestModelsIsolated(t *testing.T) {
-	s := NewInMemoryStore()
+	s := NewDatastore()
 
 	s.GetOrCreateModel("gpt-4").GetAttributes().Put("metric", testValue{Value: 1})
 	s.GetOrCreateModel("llama-3").GetAttributes().Put("metric", testValue{Value: 2})
@@ -76,7 +76,7 @@ func TestModelsIsolated(t *testing.T) {
 
 // TestIndependentStoreInstances tests that two Store instances are fully isolated.
 func TestIndependentStoreInstances(t *testing.T) {
-	s1, s2 := NewInMemoryStore(), NewInMemoryStore()
+	s1, s2 := NewDatastore(), NewDatastore()
 	s1.GetOrCreateModel("llama-3").GetAttributes().Put("key", testValue{Value: 1})
 	if _, ok := s2.GetOrCreateModel("llama-3").GetAttributes().Get("key"); ok {
 		t.Error("expected s2 to be independent from s1")
@@ -85,7 +85,7 @@ func TestIndependentStoreInstances(t *testing.T) {
 
 // TestModels tests that Models() returns all tracked model names with correct content.
 func TestModels(t *testing.T) {
-	s := NewInMemoryStore()
+	s := NewDatastore()
 	s.GetOrCreateModel("gpt-4")
 	s.GetOrCreateModel("llama-3")
 	s.GetOrCreateModel("mistral")
@@ -95,7 +95,6 @@ func TestModels(t *testing.T) {
 		t.Errorf("expected 3 models, got %d", len(models))
 	}
 
-	// Verify actual model names are present
 	expected := map[string]bool{"gpt-4": true, "llama-3": true, "mistral": true}
 	for _, name := range models {
 		if !expected[name] {
@@ -110,7 +109,7 @@ func TestModels(t *testing.T) {
 
 // TestConcurrentAccess tests thread-safety of concurrent GetOrCreateModel calls.
 func TestConcurrentAccess(t *testing.T) {
-	s := NewInMemoryStore()
+	s := NewDatastore()
 	var wg sync.WaitGroup
 
 	models := make([]datalayer.Model, 50)
@@ -132,20 +131,17 @@ func TestConcurrentAccess(t *testing.T) {
 
 // TestAttributeNilValue tests that nil values are ignored (no-op) as per AttributeMap contract.
 func TestAttributeNilValue(t *testing.T) {
-	s := NewInMemoryStore()
+	s := NewDatastore()
 	m := s.GetOrCreateModel("test-model")
 	attrs := m.GetAttributes()
 
-	// Attempt to store nil value (should be ignored)
 	attrs.Put("nil-key", nil)
 
-	// Verify key does not exist (nil values are not stored)
 	_, ok := attrs.Get("nil-key")
 	if ok {
 		t.Error("expected key to not exist when nil value is provided (nil values are ignored)")
 	}
 
-	// Verify that storing a valid value works
 	attrs.Put("valid-key", testValue{Value: 42})
 	val, ok := attrs.Get("valid-key")
 	if !ok {
@@ -158,13 +154,12 @@ func TestAttributeNilValue(t *testing.T) {
 
 // TestConcurrentAttributeAccess tests concurrent reads and writes to model attributes.
 func TestConcurrentAttributeAccess(t *testing.T) {
-	s := NewInMemoryStore()
+	s := NewDatastore()
 	m := s.GetOrCreateModel("concurrent-model")
 	attrs := m.GetAttributes()
 
 	var wg sync.WaitGroup
 
-	// 5 concurrent writers
 	for i := 0; i < 5; i++ {
 		wg.Add(1)
 		go func(id int) {
@@ -176,21 +171,19 @@ func TestConcurrentAttributeAccess(t *testing.T) {
 		}(i)
 	}
 
-	// 5 concurrent readers
 	for i := 0; i < 5; i++ {
 		wg.Add(1)
 		go func(id int) {
 			defer wg.Done()
 			key := fmt.Sprintf("key-%d", id)
 			for j := 0; j < 100; j++ {
-				attrs.Get(key) // Just read, don't care about value
+				attrs.Get(key)
 			}
 		}(i)
 	}
 
 	wg.Wait()
 
-	// Verify all keys exist
 	for i := 0; i < 5; i++ {
 		key := fmt.Sprintf("key-%d", i)
 		if _, ok := attrs.Get(key); !ok {
